requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
email-validator==2.2.0
openpyxl==3.1.5
reportlab==4.2.2
xhtml2pdf==0.2.16
